requirements.txt
streamlit>=1.37
python-dotenv>=1.0

# Optional (uncomment as needed)
# python-dotenv>=1.0
# httpx>=0.27
# pydantic>=2.8
